What is the best width (to get the best clarity) for the images to appear on the individual portfolio pages? Minimum width 1400px. That’s the width of the design. Of course, on a wider screen it will blow out a little.
Also, the text inside the bubbles on the main portfolio page is left aligned and should be centered (as in example 2 below). Can you also expand the width of the text area.

In the design the circle that appears when hovered (example 2 below) is not as blurred as the way you have done it (as shown in example 1). We would like it to look more like the design. Can you do this? I am guessing a little, but yeah, the blur can be altered.

On the Portfolio detail page, can you change Solutions to Solution and Results to Result.